got a working study definition and working measures function
worked with OpenSafely support to get a working study_definition.
Changes made so it is able to import codes from the codelist.py
modified the variable age_cat as ratio didnt add up to 1.
updated the project.yaml file to include date range in the 'gernerate_study_population in project.yaml so the measures knows the study population.